Main CRUD functions added. * ability to add, edit, remove, move all entities; * filtering items by name or/and numbers; * ascending and descending filtering; * context menu for tree; * backend simplification.

+++ src/main/java/ru/gctc/inventory/server/db/repos/ContainerRepository.java
@@ -10,7 +10,7 @@ import ru.gctc.inventory.server.db.entities.InventoryEntity;
 public interface ContainerRepository extends InventoryRepository<Container> {
     Page<Container> findAllByRoom(InventoryEntity room, Pageable pageable);
 
-    long countAllByRoom(InventoryEntity room);
+    int countAllByRoom(InventoryEntity room);
 
     boolean existsContainerByRoom(InventoryEntity room);
 }
